move NetConfig
